+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## TBA
 
+* fixed
+    * imgbabes.com
 * removed
     * adjoin.me
     * linkbee.com