+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     * dumppix.com
     * subirimagenes.com
     * screenlist.ru
+    * freeporndumpster.com
 * removed
     * adjoin.me
     * linkbee.com