+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     * screenlist.ru
     * freeporndumpster.com
     * hotimg.com
+    * freebunker.com
 * removed
     * adjoin.me
     * linkbee.com