+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     * freeporndumpster.com
     * hotimg.com
     * freebunker.com
+    * ibunker.us
 * removed
     * adjoin.me
     * linkbee.com