+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## TBA
+## 2.22.0
 
 * fixed
     * imgbabes.com